Added a Timer that Works
Timer is a separate class/thread.  It currently counts up properly and
System.out.printlns the current time elapsed when a bubble is click
(YAY).  Next, it will need to be javadoced (will do tomorrow) and be
printed on the screen.  I'm trying to make it so that a new timer object
doesn't need to be created each time a level is run.