Display time now displays the time and saves scores
the scores are correctly saved in the file, but they are not currently
calculated.  they are just put in as being the time they took.  also,
the display time page is not very pretty at all.  got to work on chm,
but it's in a local folder on my computer at the moment.